Remove unnecessary try/except from token provider
The validate_v2_token method previously had a try/except that would catch any
ValidationErrors and raise a NotFound along with logging an error message. It
turns out that there were no opportunities for a ValidationError to be raised
from within the validate_v2_token method, making the try/except unnecessary and
untested.

This patch removes the try/except completely.

Change-Id: I0277ae89c0f8551c7e795ec1331af7da06134102