Emit log message for fernet tokens only
A previous refactor improved some of the logging we had when loading Fernet
keys. The log message makes sense from a token perspective since
`max_active_keys` is configured for the fernet token provider. For the
credential provider, it doesn't necessarily make sense because we don't expose
a `max_active_keys` option through configuration.

This commit makes it so the we only emit the log message when it is referencing
a fernet token key repository.

Change-Id: Iad04bd1340a9a59c465337dedcfcd1ea0bed4bde