Consistently round down timestamps
This is one of the ways we can prevent race conditions with backends that round
datetime objects or strings before persisting them.

Closes-Bug: 1622010
(cherry picked from commit 301b6a7bc770830485937f0b9927a26e2e5ec8c8)

Conflicts:
	keystone/tests/unit/test_v3_auth.py: freezegun was added only in Newton
	keystone/tests/unit/test_v3_os_revoke.py: minor conflict

In addition to cherry-pick, time.sleep() was added to several tests.
The tests assume that some time must pass between some operations.
In Newton and later this was done in other, unrelated commits and
freezegun was used. Freezegun cannot be used in Mitaka. Because of
that, time.sleep() was added at the same places where freezegun's
tick() is used in Newton.

Change-Id: I7c6d525dfb4ec13edb360a77b27422310d545305