Remove the APIs from doc that is not supported yet
The support of updating project or deleting project cascadely has
not yet implemented in the controller [1] and no routed is defined
yet for this feature.

So, remove they from the public doc and can be added back when this
feature is fully implemented in order not to mislead readers.

[1] https://review.openstack.org/#/c/243585/

bp project-tree-deletion

Change-Id: I22b2cb41d44975bca6bc96ba2e3daa61d5029cce