Remove useless method override
The default behavior when a method is called on an object is to call
the implementation in the superclass. So no need to override a
method just to call the superclass implementation.

This makes the code less confusing because readers will wonder what
the code doing here when it doesn't do anything.

Change-Id: If2a5e694114a97d989f88c16691062dbd7215427