Wrap invalidation region to context-local cache
Since we started to use invalidation strategy, region_id of cache
is stored in cache itself in a special region. This requires fetching
region id from memcached, which might be slow. In order to prevent
unnecessary lookups of region id, wrap the code in request-local
cache, which is very fast.

This change increases performance and doesn't fix functionality
bugs.

Closes-Bug: 1641652

Change-Id: I386cf15c51f88ac4a733ec357f62ebebda8d3699