Minor fix in role_assignments api-ref
GET /role_assignments do not list the effective role_assignments,
this is done via the "effective" query param as described in the
doc.

For listing role assignments, we also have a similar explanation in
the "roles.inc" doc. After giving some thought, I guess the best option
is to leave both entries. This can sound redundant, but both entries give
pertinent explanation to the doc they are part of. For example, in the
"inherit.inc" doc, we have a introduction to the GET /role_assignments
API and explanations about the "effective", "include_subtree" and
"inherit_to" query parameters, which are essentially part of the
inherited roles feature.

Closes-Bug: 1645554

Change-Id: I38fa771295a1e1f482b10013f922a0bd0e432f8d