Default the assignment backend to SQL
Previously, keystone had logic to default the assignment backend,
without providing a default through configuration. This behavior was
removed and we should provide a sensible default for the assignment
backend.

This commit defaults the assignment backend to `sql` since that is
the only backend keystone currently supports. We also leave a warning
that there isn't any reason to change the default unless a deployer
is writing a proprietary or custom assignment backend.

Change-Id: I6fb9b7fa401c0805a538b46141a4d8c8df2093cf
Closes-Bug: 1630435