cache_on_issue default to true
If caching is globally enabled and enabled for tokens then we should
default to have cache_on_issue default to True as well. This commit also
adds one clarifying sentence to the config file comments since this
value has no effect unless caching is globally enabled and enabled in
the [token] section.

Closes-Bug: #1641816

Change-Id: I2797964ef0bb5641d8c1a208e1b5dcf567c1437f