refactor the token controller
The token controller had quite a few case statements depending on
which type of authentication was required based on the format of the
authenticate request.

This commit uses the generator pattern to give the
token.controllers.Auth.authenticate() method the right authentication
object based on the format of the request. Each authentication method
just implements an `authenticate()` method that returns a tuple of
information.

Change-Id: I11f508dd55c910b122c4102a96252041f76d6224