Fix test_revoke to run all tests after pki removal
The recent change [1] caused the removal of a certain class that
looked specific to PKI and PKIz testing and implementation, however,
actually was being used to test all supported token providers. This
caused a chunk of tests to simply not run at all and therefore be
overlooked. This change adds support for both UUID and Fernet so that
all tests are being run once again and for all available token
providers.

[1]: Icf1ebced44a675c88fb66a6c0431208ff5181574
Change-Id: I5365e2ed74ea55377729d0910cc2892d2bb8889c