Require domain_id when registering Identity Providers
An Identity Provider (IdP) should be related (1:1) to a domain so that
federated users properly belong to a domain and can be uniquely
identified by their domain + unique_id. This patch makes it so that a
domain_id is required when registering a new IdP. If not explicitly set
via the API, the IdP will be mapped to a newly created domain. The docs
and release notes will be added in a subsequent patch.

Partial-Bug: #1642687
Partially-Implements: bp support-federated-attr
Change-Id: Id18b8b2fe853b97631bc990df8188ed64a6e1275