Refactors _get_names_from_role_assignments
The method _get_names_from_role_assignments is refactored to start
building up the assignment including names from a copy of the
original assignment (without names).

That way, key/value pairs that are just copied over (such as
inheritance attribute) do not need to be analyzed anymore.

When iterating over the key/value pairs of the original assignment,
it was naming them as id_type/id_. That naming could be better, as
pointed out in reviews on [1]. They have been simply renamed to
key/value.

[1] https://review.openstack.org/#/c/380973

Change-Id: If377f046a7ae8061f88a1d3d526b564b21645bcc