Add doctor check for debug mode enabled
This patch adds a symptom test to verify that debug mode is not
enabled. Doing so will eliminate the need to add an additional
check for debug enabled in other symptom checks like caching and
database. This also corrects the unit tests that have already been
added as well as adds a unit test for the new debug symptom.

Related-Bug: 1641621

Change-Id: I76750ab3553026583b9fb330bd39f2a134d9c230