Fixed multiple warnings in tox -edocs
There are multiple tracebacks and warnings being thrown whenever
the keystone docs are built due to documents being moved/deleted as
well as formatting issues in a couple places.

This fixes a few of the warnings due to broken links and fixes a few
of the method docs.

Master: http://paste.openstack.org/show/591730/
This Patch: http://paste.openstack.org/show/591735/

Change-Id: I11cbbc7a10fa24dcbf67c76e3061a39a58529c06
Partial-Bug: #1602422