Make bootstrap idempotent when it needs to be
This commit makes `keystone-manage bootstrap` completely idempotent
when configuration values or environment variables haven't changed
between runs. If they have changed, then `bootstrap` shouldn't be
as idempotent becuase it's changing the state of the deployment.

This commit addresses these issues and adds tests to ensure the
proper behavior is tested.

Conflicts:
    keystone/tests/unit/test_cli.py

As of Newton and newers releases, a context is no longer passed into
the controller, but a request object.

Change-Id: I053b27e881f5bb67db1ace01e6d06aead10b1e47
Closes-Bug: 1647800
  (cherry picked from commit 90f2f96e69b8bfd5058628b50c9f0083e3f293e9)