Cascade delete federated_user fk
The bug was caused by a foreign key in the federated_user table. This
key prevents a protocol from being deleted after a successful
authentication has happened (so the creation of a federated user
via shadowing). We take advantage of the same foreign key by adding the
cascade delete behavior to it.

Closes-Bug: 1642692

Change-Id: I3b3e265d20f0cfe0ee10c6a274d9bdf4e840b742