Remove stevedore warning when building docs
When running "tox -edocs", the following warning is raised:

WARNING stevedore.named [-] Could not load keystone.notifications

It looks like the removed line "keystone.notifications" is something
that was left over from another change [1] and never removed from the
config-generator. This change removes this line and the warning no
longer appears as a result.

[1] I1369fd7835b31e57a094d07deb6a1aacc1d314b5

Change-Id: I08325abadba94efdb5fc405f859bdcb766914782
Partial-Bug: #1602422