Code-Defined Resource-specific Options
Implement the code-defined resource specific options framework
for the Identity subsystem (User). Options are defined in code
in the keystone.identity.backends.resource_options. These options
are extracted from the update dict, validated for a specific type,
and stored in an attribute association (as a one[User] to
Many[one-per-option-type]) relation in a database table.

On retrival options are converted from the association to elements
in the user reference dict.

Setting an option explicitly to None removes the option (unsets it).

Updating a specific option will have no impact on other options.

Change-Id: I90875ce2c3cf899557bb1f215fa1bda9576cbe83