Cleanup for resource-specific options
This patch adds in the missing helper function in the sql_model for
user to get an option out of _resource_option_mapper by the ID
eliminating the need to access a "private" attribute directly outside
of the model itself (in most cases).

This also cleans up the USER_OPTIONS_REGISTRY list allowing direct
access to the option constants in keystone.identity.backends.resouce_options.

Change-Id: I385871499da191ea58acc4ddcafebbd94b10a489