Add MFA Rules and Enabled User options
Add the user-options for the MFA rules and the MFA Rules enabled. This
validates and stores the option values for the MFA rule sets (list of
lists of strings, no duplicate sub-lists, no duplicate strings in sub
lists).

This does not implement validation that the strings are in-fact
auth-methods. This is just the scafolding for the user-options.

Change-Id: I64add63d9fc99bf66936913523340ee3c1290a9e
bp: per-user-auth-plugin-reqs