Give a prospective removal date for all v2 APIs
All the v2 APIs are marked as deprecated, but the v2_auth and v2_ec2
did not have proposed removal dates. In keeping with the general
standards, the earliest we can do this is Pike + 4, although we
specify this as Mitaka + 7 so as not to change the actual release
version we first marked these as deprecated.

Change-Id: I1d8c5e43ff7c8e4256639324db47a64635e217df
Implements: bp removed-as-of-pike