Fix duplicate handling for user-specified IDs
For resources such as federation protocols and federation mappings, the
database primary keys are ID strings specified by the user creating
them. If the user created such a resource that happened to have the
substrings 'id' or 'name' in the identifier, and then by accident tried
to create it again, it would fail with a message that did not appear to
relate to the entry being a duplicate:

 string indices must be integers (HTTP 400)

This was because the method that is supposed to form a user-friendly
message receives all the arguments as a tuple and iterates over it,
looking for a dictionary with the keys 'id' or 'name' to figure out what
was trying to be duplicated. However, it can't distinguish between a
dictionary with 'id' or 'name' as a key and a string with 'id' or 'name'
as a substring, and trips if it finds such a string. This logic for
looking for 'id', 'name', or 'domain_id' in an object really only makes
sense if the object is a dict, so this patch adds a check to ensure it
is a dict before looking for keys in it.

Change-Id: If3c23a28eb5594efaa49c6a15d8db11cfc8d9057
Closes-bug: #1668563