Support endpoint updates in bootstrap
The current bootstrap command does not update the endpoints, but
does update many other settings.

It's very convenient to have a way to update endpoints from the
bootstrap command. One such use case is when bootstrap is run from
a configuration management tool such as ansible, and the endpoint
URLs need to be updated.

Change-Id: I2b1cb53701c3e5fb2d91c14d9c792beadd1dbbc2