Fix description for 204 response
In addition to DELETE requests, keystone may return an HTTP 204 as a
success code for PUT[1] and HEAD[2] requests. It would be quite
disconcerting if keystone were actually deleting resources during such
additive and read-only operations, so this patch simplifies the 204
description so that it is truthful in all of its use cases.

[1] https://developer.openstack.org/api-ref/identity/v3/index.html?expanded=assign-role-to-group-on-domain-detail#assign-role-to-group-on-domain
[2] https://developer.openstack.org/api-ref/identity/v3/index.html?expanded=check-whether-user-belongs-to-group-detail#check-whether-user-belongs-to-group

Change-Id: I0e4f9df873a831d7329e71b77818bc47d3c2bbb2