Speed up check_user_in_group for LDAP users
Without this patch, the HEAD /v3/groups/{group_id}/users/{user_id} call
uses the list_users_in_group to look up members of a group, which in
addition to getting the IDs of each member also queries LDAP to make
sure those users are really in LDAP. For a group with many users, this
is very slow, and also quite roundabout when we already know the ID of
the user we are looking up. This patch enhances performance by splitting
up check_user_in_group from list_users_in_group (in this test the group
contained about 1000 users):

  http://paste.openstack.org/show/603695/

Notably the performance of this patch when ldappool is enabled is about
the same as when it is not because there are fewer connections being
made.

Since there are no list actions being added this should still be safe
and accurate when list limits or paging is enabled.

Change-Id: I11b0dff1ebba4f5fca67e5329bdd1b5124cd37