Remove unused revoke_by_domain_role_assignment
This patch removes a method that wasn't being used anymore anywhere
except for a single unit test. In addition, we no longer need to store
the revocation event when the token provider API will rebuild the
token context at validation time. Some of the test coverage for this
behavior can be located here:
https://github.com/openstack/keystone/blob/master/keystone/tests/unit/test_v3_assignment.py#L1175-L1177
https://github.com/openstack/keystone/blob/master/keystone/tests/unit/test_v3_assignment.py#L1191-L1193
https://github.com/openstack/keystone/blob/master/keystone/tests/unit/test_v3_assignment.py#L1311

Change-Id: I1ee53f15ec6b2dae10bfbd0fc3435e018f26f04b
partial-bug: 1671887