Updated scope parameter description in v3 API-ref
The description of scope parameter in the API docs is inconsistent for
scoped and explicit unscoped authorization. It doesn't explain clearly
when the parameter should be set to keyword 'unscoped' (explicit
unscoped authorization) and when to include ID/name of the project/
domain (scoped authorization). This patch covers the required update
for the scope parameter by introducing separate descriptions for the
above mentioned cases.

For explicit unscoped authorization, the keystone server specifies
scope value as 'unscoped' whereas in some old versions of keystoneauth
the value is {'unscoped': {}}. Added a comment to keystone/auth/schema.py
to make sure the type of scope parameter is object and not string, so
that it is consistent with older versions too.

In v3 API docs the case for token authentication with explicit unscoped
authorization is missing. Added documentation and request example for
that case.

Partial-Bug #1637682

Change-Id: I10d38e4cc59934c421443322b14f4f971acf3a29