Remove unused revocation check in revoke_models
This patch addresses [1] by removing the corresponding check in
revocation models. There is no longer a need to check a token against
the revocation table for access_token when we no longer persist them
in a revocation event. This is due to token providers handling this
when rebuilding a token at validation time.

[1]: I4dc766981a29b0afd0a44718c1c5d81155163982
partial-bug: 1671887

Change-Id: I5a50ab9cdca64005e1e4a6738ee6a8accf458ed8