Remove usage of enforce_type
The usage of `enforce_type` from oslo.config will be removed in the
4.0 release of oslo.config. The default behavior has been
incorporated into set_override/set_default as of:

  Ifa552de0a994e40388cbc9f7dbaa55700ca276b0

We no longer need to specify `enforce_type=True` and we should
remove it since it will be removed from oslo.config.

This commit also fixes violations with enforce_type=True.

Change-Id: I8222e84583aaa4de4c7c36ec8cec5e35c2e7e253
Related-Bug: 1517839