Writing API & Scenario Tests docs
It is nice to see a real example of a real test when trying to write a
new one. This patch adds a doc that gives an overview of writing new API
and Scenario tests for the keystone tempest plugin.

This also fixes a typo in the IdentityProdiversTest class name.

Change-Id: Icd631124b451f17df3d4de184e73f6223fdb7de3