Expose a bug in domain creation from idps
When creating an identity provider, a domain will be created with it
if it isn't already provided. If a database conflict occurs when an
identity provider is created, the domain associated with it isn't
cleaned up. This essentially orphans a domain that shouldn't have
been created because the identity provider was never successfully
created.

Change-Id: Ie59d21abda422d4e9668725de4604ab99701dc59
Related-Bug: 1688188