Fix misnamed variable in config
The `caching` configuration option for the catalog section was
instead storing the configuration in a `aching` variable. I'm
assuming that this was suppose to be `caching` given the convention
of all the other sub-systems within keystone and how they implement
caching. This could have had performance impacts if operators thought
they were enabling caching behavior but actually weren't seeing any
performance increase because the implemention looks for `caching` as
the registered configuration option instead of `aching`. This seems
to have been introduced in:

  https://github.com/openstack/keystone/commit/d9c6b50a3ae514e640fa13a344e59fe3649ee0ef

Performance impact here could be significant if the catalog is of
substantial size since it would be rebuilt every time a token is
validated.

Change-Id: I04e71ba12a86106bc2e50daa139666bd89a6854c