Remove assertRaisesRegexp testing function
This change removes the function assertRaisesRegexp from within
keystone to help reduce any confusion about what is really being
called within the appropriate test. The test was changed to use
assertRaisesRegex() which is aliased in six for python 2 and 3.

assertRaisesRegexp in this change was a part of keystone itself
rather than the function of the same name from python 2, and was
causing confusion with recent changed regarding deprecation
and python 3 changing the name to assertRaisesRegex.

Change-Id: I63c84ff432e08866253cfb14ad3bb8db4a665589