A simple fix about explicit unscoped string
If a user has a default project defined in their user reference, but
does not have authorization to the default project, the default project
is ignored at token creation, which will return an unscoped response
without any authorization.
We improve the wording about explicit unscoped to make it more correct.

Change-Id: I6748dfd8aae57b1227d6287030aff3d92fe72b86