Add HEAD API to domain config
The domain configuration API had several GET calls that didn't
support HEAD. The documentation for the added HEAD APIs is going
to be proposed at a later date as a blanket statement saying all
APIs the support GET also support HEAD. This reduces duplication
of documentation for a relatively straight-forward concept.

This commit also fixes some inconsistencies in the domain
configuration documentation with colons and newlines.

Change-Id: I7a6ab2f05600c8d562d5be83d651ff735c9a68a4
Partial-Bug: 1696574