Added configuration references to documentation
Currently some of the config references docs are a part of
general OpenStack-manuals. Migrating those docs to keystone
documentation so that they can be reviewed effectively by
keystone developers too.

Following the specs, the files are added to admin/ and
configuration/ directories.

The files containing congiguration options are added
in a subsequent patch using oslo.config plugin.

Partial-Bug #1694460

Change-Id: I9a85f610e66a10dac54c50b2a54305e979888ee5