Improve handling of database migration checks
The `--check` subcommand is suppose to provide useful information
and status codes depending on the state of the keystone database.
Operators and automation use this information to determine what their
next step is in a rolling upgrade. The current logic is broken
becuase it doesn't account for new installations that might be
relying on this information.

This change breaks that case into multiple try/except statements and
handles each appropriately so that the status code and logging
information is accurate for operators and automation using this
information for upgrading a new keystone database.

Change-Id: I331fa663a99f79ea9a79a75e4ae07c45278556bf
Closes-Bug: 1698900