Remove the revocation api config section
In parsing the configuration.rst doc and porting important bits over
to the admin-guide, I noticed the revocation events section isn't
very accurate.

The revocation API (GET /v3/OS-REVOKE/events) has been deprecated and
only returns an empty list, since it was only used for PKI/Z tokens,
which are no longer supported by keystone.

Otherwise revocation events are a structure that keystone only uses
internally. This commit removes it from the configuration.rst guide.
The information that was accurate in the guide is already documented
in configuration and the API.

Change-Id: I45d16fd50239efc543fe20b228b021f65cfd104d