Remove keystone_tempest_plugin from setup.cfg
Since keystone_tempest_plugin is moved to a seperate repo, but we
have not remove its entry from packages section in setup.cfg. On
Installing keystone, It also installs keystone_tempest_plugin in
python sitelib directly which contains README.rst file only. which
does not addes value. So it is better to remove that entry.

Change-Id: I832600d841e9a84c3dfee7b5c3e8893ece1ae902