Use log debug instead of warning
Setting the default project is merely provided for convenience,
does not need exist project and grant any actual authorization
on it, This shouldn't be a warning.

At the same time, if the log message exposes to the users,
they know the identity authentication has already passed and
the used password is valid, which could be useful to an attacker.
Log warnings is not necessarily as locked down as debug log,
I think this would be fine as a debug log.

Change-Id: I85ed00ad3d08ed52b159d4da1526a006d72d9677