remove default rule
The default rule no longer applies with the move of policy into code
so this change removes it.

In previous releases, the default rule was used by operators
customizing policy and not wanting to specify every rule in their
policy.json. But with the move of policy into code, all checks that
the code is going to make are defined in code, so there should never
be an occasion for the default rule to be checked. Leaving it defined
would confuse operators since it can no longer be used the way it was
used before.

Change-Id: Idafe1c906f1eb188200eab7af3eae8eb86c8154a
Closes-Bug: #1703392