Improved the keystone federation image
The image migrated as a part of openstack-manuals is
a little pixelated and blurred. This commit replaces
the image with a higher resolution one.

Also a .svg version of this file is updated. The file
is not used as such in the docs but is useful if
someone wants to make changes to the png image, which
can be generated from the .svg file.

Change-Id: I56b863974adb5642f47e7561aa70bc582c85485f