Add database migration for project tags
This change adds the SQL database migration for adding a new
project_tag table that consists of an project_id and name. The
project_id is the id of a project that the tag is mapped to and
the name is the string value of the tag.

Partially-Implements: bp project-tags
Co-Authored-By: Pilla, Samuel (sp516w) <sp516w@att.com>
Co-Authored-By: Tin Lam <tin@irrational.io>

Change-Id: Ib3db74cb08e7fdec01d55d8988e0906948d80a32