Filter users and groups in ldap
All users and groups are required to have a name. Prior to this fix,
Keystone was not properly enforcing this for LDAP users and groups.
Keystone will now ignore users and groups that do not have a value for
the LDAP attribute which Keystone has been configured to use for that
entity's name.

Closes-Bug: #1704205

Change-Id: I424813785b7a355514ef42f1e4c6384b8a78b256