Emit deprecation warning for federated domain/project APIs
Before the identity API documentation lived in-tree, it lived in the
openstack/keystone-specs project. Before that it lived in it's own
identity-api repository. A long time ago we merged a patch [0] to that
project to deprecate the project and domain APIs in the federation
path. That documentation has since been copied and migrated to the
keystone-specs repository and now keystone itself.

Even though we've deprecated those API in the documentation, we never
formally deprecated them in code using ``versionutils``. This commit
does that. Now a message will be logged anytime that API is used
instead of GET /auth/domains or GET /auth/projects. This makes the
implementation of keystone consistent with documentation we merged
ages ago.

[0] If0c010119512d6a159ce82147a87c4698205a648

Change-Id: I88cb587c7cd69fef400b2eaf7dffda655355b710