Add user/ directory for documentation
The specification detailing the great documentation migration
requires a doc/source/user/ directory be added to the project.
The directory is intended for end user documentation.

This commit add the directory so that we are in compliance with the
specification but doesn't move any content into the directory. It is
instead treated like a placeholder until we can migrate the right
content into the user guide.

Change-Id: I7d9cde958957bff20c7221f37cb0b03b61aa38f9
Partial-Bug: 1694460