Clarify SELinux note in LDAP documentation
The eye-catchy-ness of the Important SELinux notice in the LDAP
integration documentation can cause confusion for users to whom the
notice is not applicable, such as Ubuntu or SUSE users. For them,
SELinux is not enabled by default and they may not even be aware of what
it is for, and so perceive a failure when they try to enable the SELinux
boolean as described in the document. This patch prepends a
clarification that the notice only applies to SELinux users, to
hopefully make it clearer that not being able to follow the directions
is just fine in many cases.

Change-Id: I65a34608f1a51ec923329065b6443bdd525a1ac7