Add description for relationship links in api-ref
This adds a section within the index file that describes what a
relationship link is and what it is used for in terms of each
operation within keystone. There will be a relationships section
in both v3 and v3-ext.

This should help clarify any confusion that may arise when a user is
viewing the api-ref about what the relationship links are.

Change-Id: I9c6b7959ed6ea682c565c515af0cf509b6a64e5d
Closes-Bug: #1674676