Add explain of mapping group attribute
As we use federation protocol, assign authorization to ephemeral
users by mapping to groups. Setting this mapping does not add the
ephemeral user to the group and the user is not a member of the group.
Only authorization for the user is same as role assignments to the
group on a project or domain.

This patch add explain about the group mapping

Change-Id: I9faa7d57037af3c2cf6ccfda8d853693fa5eb628