Ensure building scope is mutually exclusive
The authentication tests have a helper method that builds
authentication requests based on parameters. The logic in this method
made is possible to generate scopes that didn't make sense (e.g.
domain scope and project scope in the same request). This commit
uses else/if statements to ensure only one valid scope is represented
in the request.

Change-Id: Ie7e156a83006c1578c87d862cb4d1d948a800809