Copy specific distro pages for install guide
This change breaks the get-started-identity, keystone-users,
and keystone-openrc pages into multiple copies for each distribution
release in the keystone install guides. Before this change, sphinx
would not render the links correctly and these pages would be skipped
in both the RDO and Ubuntu guides.

To avoid unneccessary duplication of the same text, the files were
moved into common/ and are simply included into the newly created
files for each release.

Change-Id: I7b478504f1e68034ea993c0214705045da86bb6e
Closes-Bug: #1712335