Remove v2.0 auth APIs
This was originally staged to be removed in the T release.
Discussions from the Queens PTG resulted in the ability to remove it
sooner since everything else in v2.0 is gone except the ec2 APIs.

This patch just removes the v2.0 authentication API and the tests
that failed as a result. A subsequent patch will go through and start
removing all the plumbing, fixtures, and testing infrastructure that
is no longer needed.

bp removed-as-of-queens

Change-Id: I4c3e35f3565b4b60ae4d00cc2490bd04aba1a800