Add a new table for system role assignments
This commit introduces an additive only database migration for a new
database table. This table will be used only for system-level role
assignments instead of tacking the functionality into the existing
assignment table. The reason for this is that it will be easier to
implement as a separate table and it will be easier to extend
functionality in the future to support more complex system role
assignments.

bp system-scope

Change-Id: Ia0722004b7b851c46e8a72780b3acd31f16f78a6