Use stestr directly instead of ostestr
This is inline with the direction from:

  http://lists.openstack.org/pipermail/openstack-dev/2017-September/122135.html

This also makes it easier to invoke tests based on a regular
expression. Previously, it was possible to do:

  $ tox -e py27,py34 -- keystone.tests.unit.test_module.TestCase

But that isn't supported in the way we use ostestr in tox with the
new stestr changes. Instead of refactoring our ostestr usage we can
just replace it with stestr directly which allows us to invoke tests
using a single regex and is inline with the direction of moving away
from ostestr.

Change-Id: I83eb6ce905c45c9c388a8d5a3c9369f5c5f6710e