Fix initial mapping example
The first example of federation mapping rules that appears in the
Configure Federation docs is incorrect. The "any_one_of" matcher returns
a boolean, not the value of the match, and so it would not map to a
local rule like "name": "{0}". This patch fixes the example so that it
works properly but still has the intended meaning, which is to filter on
username.

Change-Id: I1f69f5cdde470811045fc60a72e50c2bd688be53