Add non-voting rolling upgrade test
This patch implements a replacement of the legacy rolling
upgrade test previously implemented. The job and pipeline
is defined in-repo to ensure that the keystone team has
complete control over the voting status, pipelines it is
in, etc.

The test uses the 'upgrade' tox environment in the
openstack/openstack-ansible-os_keystone repository which
executes a build of two keystone containers using the
the head of the stable branch of keystone, then upgrades
to the proposed review. The upgrade is a complete rolling
upgrade using MariaDB as a database.

Depends-On: If39e4bec8db0be04b13dfc005ee404b73c0b65d3
Depends-On: Ia5ab6aa43f912b3cb4adf98154cd3da12688c2e0
Change-Id: I78bb5e18c1ce0ddf18494f7cd0ef1ab6953be771