Remove the sample .conf file
The sample configuration file is almost never up to date as it has to
be updated by a person submitting a patch.

The implementation of published autogenerated sample config files in
documentation was done in I88a2429dd3cacd1d014b5b441b98fbfee7e1e208
and in If00cd3bcc654a45944c0bc8b3f146c75bd970f9a. These generate
sample configuration files and publish them in the documentation on
every commit, ensuring that they are always up to date, and not
requiring human intervention to be updated.

As has been done with nova (in Mitaka), cinder (in Newton), and
neutron (in Newton) this patch removes the sample config file from the
git tree and replaces it with a README file explaining how to generate
them, or where to find the latest published versions in the online
documentation.

This commit also breaks a related testcase into two distinct tests for
easier readability, making it clearer what the behavior is through one
assertion/concept per test.

Depends-On: https://review.openstack.org/#/c/562007/
Change-Id: Ic4d6a98035f59b6ebe48d9c85af50fc9408fc3ab