Expose a bug when authenticating for a trust-scoped token
The test added here should pass with a 400 Bad Request, but keystone
doesn't validate the OS-TRUST:trust key of the authentication request
is actually a dictionary. As a result, it assumes it is a dictionary
and attempts to use dictionary methods on. This results in a 500
Internal Server Error when it should really be a 400 Bad Request.

A subsequent patch can fix this by enhancing the jsonschema
definition for token authentication.

Change-Id: I9ea118945870f27dd73000b45673d966ea05ca53
Partial-Bug: #1733754